management: Move inactive_since to digest.
